+++ ebi/ols/api/helpers.py
@@ -12,13 +12,16 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 """
+import logging
 import re
 from collections import namedtuple, OrderedDict
 
 import inflection
 from coreapi import Object
 
-from ebi.ols.api.base import *
+from ebi.ols.api.base import ListClientMixin, site
+
+logger = logging.getLogger(__name__)
 
 
 def underscore(value):
@@ -28,7 +31,7 @@ def underscore(value):
 
 def convert_keys(data):
     """
-    Convert Json like keyx into Python-like keys
+    Convert Json like keys into Python-like keys
     :param data: a OrderedDict or None
     :return: OrderedDict
     """
@@ -51,6 +54,63 @@ def to_python_value(value):
     return value
 
 
+class HasAccessionMixin(object):
+    short_form = None
+    obo_id = None
+    is_obsolete = False
+    is_defining_ontology = False
+    iri = None
+
+    _accession = None
+
+    @property
+    def accession_old(self):
+        """
+        Return term accession issued from related retrieved data, try to figure out related obo_id if value not
+        set in OLS
+        :return: str
+        """
+        if not self.obo_id and self.short_form:
+            log = logging.getLogger('ols_errors')
+            log.error('[NO_OBO_ID][%s][%s]', self.short_form, self.iri)
+            # guess
+            sp = self.short_form.split('_')
+            if len(sp) == 2:
+                self.obo_id = ':'.join(sp)
+                return self.obo_id
+            else:
+                # no '_' character in short_form might ignore the error (may be #Thing)
+                logger.warning('Unable to parse %s', self.short_form) if len(sp) == 1 else None
+                return False
+        return self.obo_id
+
+    @property
+    def accession(self):
+        if self._accession:
+            return self._accession
+        if not self.obo_id:
+            logger.error('[NO_OBO_ID][%s][%s]', self.short_form, self.iri)
+            to_parse = self.short_form or self.iri.rsplit('/', 1)[-1]
+            # guess
+            sp = to_parse.split('_')
+            if len(sp) >= 2:
+                left = '_'.join(sp[:-1])
+                right = sp[len(sp) - 1]
+                accession = ':'.join([left, right])
+                logger.debug('Accession sorted out %s', accession)
+                self._accession = accession
+                return accession
+            else:
+                # no '_' character in short_form might ignore the error (may be #Thing)
+                logger.warning('Unable to parse %s', self.short_form) if len(sp) == 1 else None
+                return None
+        return self.obo_id
+
+    @accession.setter
+    def accession(self, accession):
+        self._accession = accession
+
+
 class OLSHelper(object):
     """
     Base Transfer object, mainly assign dynamically received dict keys to object attributes
@@ -58,9 +118,7 @@ class OLSHelper(object):
 
     def __init__(self, **kwargs):
         converted = convert_keys(kwargs)
-        for name, value in converted.items():
-            # print(name, value, to_python_value(value))
-            self.__setattr__(name, to_python_value(value))
+        [self.__setattr__(name, to_python_value(value)) for name, value in converted.items()]
 
     def __eq__(self, other):
         if isinstance(other, self.__class__):
@@ -123,7 +181,6 @@ class Ontology(OLSHelper):
     An Ontology element from EBI Ontology lookup service
     """
     path = 'ontologies'
-
     ontology_id = None
     loaded = None
     updated = None
@@ -146,17 +203,17 @@ class Ontology(OLSHelper):
 
     def terms(self, filters=None):
         """ Links to ontology associated terms"""
-        client = ListClientMixin('ontologies/' + self.ontology_id, Term)
+        client = ListClientMixin('/'.join([site, 'ontologies/' + self.ontology_id]), Term)
         return client(filters=filters)
 
     def individuals(self, filters=None):
         """ Links to ontology associated individuals """
-        client = ListClientMixin('ontologies/' + self.ontology_id, Individual)
+        client = ListClientMixin('/'.join([site, 'ontologies/' + self.ontology_id]), Individual)
         return client(filters=filters)
 
     def properties(self, filters=None):
         """ Links to ontology associated properties"""
-        client = ListClientMixin('ontologies/' + self.ontology_id, Property)
+        client = ListClientMixin('/'.join([site, 'ontologies/' + self.ontology_id]), Property)
         return client(filters=filters)
 
     @property
@@ -183,18 +240,6 @@ class Ontology(OLSHelper):
             self.config.namespace = namespace
 
 
-class OboXref(OLSHelper):
-    database = None
-    id = None
-    description = None
-    ur = None
-
-
-class OboCitation(OLSHelper):
-    definition = None
-    oboXref = None
-
-
 class TermAnnotation(OLSHelper):
     database_cross_reference = None
     has_obo_namespace = []
@@ -212,22 +257,9 @@ class TermAnnotation(OLSHelper):
             self.id, self.has_obo_namespace, self.alternative_term, self.has_alternative_id)
 
 
-class TermsLink(OLSHelper):
-    parents = None
-    ancestors = None
-    hierarchicalParents = None
-    hierarchicalAncestors = None
-    jstree = None
-    children = None
-    descendants = None
-    hierarchicalChildren = None
-    hierarchicalDescendants = None
-    graph = None
-
-
-class Term(OLSHelper):
+class Term(OLSHelper, HasAccessionMixin):
     path = 'terms'
-    iri = None
+
     label = None
     _description = []
     annotation = None
@@ -235,13 +267,9 @@ class Term(OLSHelper):
     ontology_name = None
     ontology_prefix = None
     ontology_iri = None
-    is_obsolete = False
     term_replaced_by = None
-    is_defining_ontology = False
     has_children = False
     is_root = False
-    short_form = None
-    obo_id = None
     in_subset = None
     obo_definition_citation = None
     obo_xref = None
@@ -259,46 +287,21 @@ class Term(OLSHelper):
     @property
     def relations_types(self):
         if self._relations_types is None:
-            client = ListClientMixin('ontologies/' + self.ontology_name + '/terms/' + uri_terms(self.iri), Term)
-            self._relations_types = [name for name in client.document.links.keys() if
-                                     name not in ('graph', 'jstree', 'descendants', 'ancestors', 'hierarchicalParents',
-                                                  'hierarchicalAncestors', 'hierarchicalChildren',
-                                                  'hierarchicalDescendants')]
+            client = ListClientMixin(
+                site + '/ontologies/' + self.ontology_name + '/terms/' + ListClientMixin.make_uri(self.iri), Term)
+            self._relations_types = [name for name in client.document.links.keys() if name not in ('graph', 'jstree')]
         return self._relations_types
 
     def load_relation(self, relation):
-        client = ListClientMixin('ontologies/' + self.ontology_name + '/terms/' + uri_terms(self.iri), Term)
+        client = ListClientMixin(site + '/ontologies/' + self.ontology_name + '/terms/' + ListClientMixin.make_uri(self.iri),
+                                 Term)
         return client(action=relation)
 
-    def children(self):
-        return self.load_relation('children') if self.has_children else []
-
-    def descendants(self):
-        return self.load_relation('descendants')
-
-    def ancestors(self):
-        return self.load_relation('ancestors')
-
-    def parents(self):
-        return self.load_relation('parents')
-
-    def hierarchical_parents(self):
-        return self.load_relation('hierarchicalParents')
-
-    def hierarchical_ancestors(self):
-        return self.load_relation('hierarchicalAncestors')
-
-    def hierarchical_children(self):
-        return self.load_relation('hierarchicalChildren')
-
-    def hierarchical_descendants(self):
-        return self.load_relation('hierarchicalDescendants')
-
     def graph(self):
-        return self.load_relation('graph')
+        raise NotImplementedError('This link process is not implemented yet')
 
     def jstree(self):
-        return self.load_relation('jstree')
+        raise NotImplementedError('This link process is not implemented yet')
 
     @property
     def obo_name_space(self):
@@ -320,12 +323,9 @@ class Term(OLSHelper):
 
     @property
     def subsets(self):
+        """ Returned coma separated list of term subsets """
         return ','.join(
             [subset for subset in sorted(self.in_subset, key=lambda s: s.lower())]) if self.in_subset else ''
-
-    @property
-    def accession(self):
-        return self.obo_id
 
     @property
     def name(self):
@@ -335,41 +335,31 @@ class Term(OLSHelper):
 Subset = namedtuple("Subset", ["terms"])
 
 
-class Individual(OLSHelper):
+class Individual(OLSHelper, HasAccessionMixin):
     path = 'individuals'
-    iri = None
     label = None
     description = []
     synonyms = None
     ontology_name = None
     ontology_prefix = None
     ontology_iri = None
-    is_obsoloete = None
-    is_defining_ontology = None
-    short_form = None
-    obo_id = None
 
     def __repr__(self):
-        return '<Individual(label={}, iri={}, ontology_name={}, short_form={}, obo_id={})>'.format(
+        return '<Individual(label={}, iri={}, ontology_name={}, short_form={}, accession={})>'.format(
             self.label, self.ontology_name, self.iri, self.short_form, self.obo_id)
 
 
-class Property(OLSHelper):
+class Property(OLSHelper, HasAccessionMixin):
     path = 'properties'
     annotation = None
     synonyms = None
-    iri = None
     label = None
     description = None
     ontology_name = None
     ontology_prefix = None
     ontology_iri = None
-    is_obsolete = None
-    is_defining_ontology = None
     has_children = None
     is_root = None
-    short_form = None
-    obo_id = None
 
     def __repr__(self):
         return '<Property(label={}, iri={}, ontology_name={}, short_form={}, obo_id={})>'.format(