+++ tests/test_basic.py
@@ -12,16 +12,19 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 """
+import logging
 import unittest
 import warnings
-import os
-import sys
 
+import ebi.ols.api.exceptions
 import ebi.ols.api.helpers as helpers
-from ebi.ols.api.base import ListClientMixin
 from ebi.ols.api.client import OlsClient
 
-os.environ["PYTHONWARNINGS"] = "ignore"
+logging.basicConfig(level=logging.DEBUG,
+                    format='%(asctime)s %(levelname)s : %(name)s.%(funcName)s(%(lineno)d) - %(message)s',
+                    datefmt='%m-%d %H:%M:%S')
+
+logger = logging.getLogger(__name__)
 
 
 def ignore_warnings(test_func):
@@ -41,37 +44,34 @@ class OntologyTestSuite(unittest.TestCase):
         self.assertIsInstance(ontology.config, helpers.OntologyConfig)
         self.assertIsInstance(ontology.config.annotations, helpers.OntologyAnnotation)
 
-    def _checkProperty(self, property):
-        self.assertIsInstance(property, helpers.Property)
+    def _checkProperty(self, prop):
+        self.assertIsInstance(prop, helpers.Property)
 
     def _checkProperties(self, properties):
-        self.assertEqual(properties.page_size, 1000)
-        [self._checkProperty(property) for property in properties]
+        if type(properties) is helpers.Property:
+            properties = [properties]
+        [self._checkProperty(prop) for prop in properties]
 
     def _checkIndividual(self, individual):
         self.assertIsInstance(individual, helpers.Individual)
 
     def _checkIndividuals(self, individuals):
-        self.assertEqual(individuals.page_size, 1000)
         [self._checkIndividual(ind) for ind in individuals]
 
     def _checkOntologies(self, ontologies):
-        self.assertEqual(ontologies.page_size, 1000)
         [self._checkOntology(ontology) for ontology in ontologies]
 
     def _checkMixed(self, helper):
         return getattr(self, '_check' + helper.__class__.__name__)(helper)
 
-    @ignore_warnings
     def setUp(self):
-        super().setUp()
-        self.client = OlsClient()
+        warnings.simplefilter("ignore", ResourceWarning)
+        self.client = OlsClient(page_size=100)
 
-    @ignore_warnings
     def test_ontologies_list(self):
         # standard first page
         ontologies = self.client.ontologies()
-        self.assertEqual(ontologies.page_size, 1000)
+        self.assertEqual(ontologies.page_size, self.client.page_size)
         total_pages = ontologies.pages
         current = 0
         num_pages = 0
@@ -85,12 +85,11 @@ class OntologyTestSuite(unittest.TestCase):
                 num_pages += 1
         item = ontologies[185]
         self.assertEqual(test_item, item)
-        self.assertEqual(ontologies.page, 0)
+        self.assertEqual(ontologies.page, len(ontologies) // ontologies.page_size - 1)
         self.assertEqual(total_pages, num_pages)
         self.assertEqual(current, len(ontologies))
         self.assertEqual(num_pages, ontologies.pages)
 
-    @ignore_warnings
     def test_ontology(self):
         # FIXME add further testing on single ontology data
         ontology = self.client.ontology('aero')
@@ -102,10 +101,8 @@ class OntologyTestSuite(unittest.TestCase):
         self.assertTrue(hasattr(term, 'obo_xref'))
 
     def _checkTerms(self, terms):
-        self.assertEqual(terms.page_size, 1000)
         [self._checkTerm(term) for term in terms]
 
-    @ignore_warnings
     def test_ontology_terms(self):
         """
         Tess retrieve ontology terms
@@ -114,42 +111,46 @@ class OntologyTestSuite(unittest.TestCase):
         # Search for all terms in ontology, loop over and load Termn accordingly
         ontology = self.client.ontology("aero")
         terms = ontology.terms()
-        self.assertEqual(terms.page_size, 1000)
         self.assertEqual(terms.index, 0)
         self.assertGreaterEqual(len(terms), ontology.number_of_terms)
         self._checkTerms(terms)
 
-    @ignore_warnings
     def test_ontology_individuals(self):
         ontology = self.client.ontology("aero")
         individuals = ontology.individuals()
         self.assertGreaterEqual(len(individuals), ontology.number_of_individuals)
         self._checkIndividuals(individuals)
 
-    @ignore_warnings
     def test_ontology_properties(self):
         ontology = self.client.ontology("aero")
         properties = ontology.properties()
         self.assertGreaterEqual(len(properties), ontology.number_of_properties)
         self._checkProperties(properties)
 
-    @ignore_warnings
     def test_list_range(self):
         ontology = self.client.ontology("aero")
-        terms = ontology.terms()
-        slice_terms = terms[3:50]
-        self.assertEqual(47, len(slice_terms))
-        self.assertEqual(slice_terms[0], terms[3])
-        self.assertEqual(slice_terms[len(slice_terms) - 1], terms[49])
+        terms = ontology.terms({'size': self.client.page_size})
 
-    @ignore_warnings
-    def test_ontology_terms_filters(self):
+        slice_terms = terms[23:253]
+        term_3 = terms[252]
+
+        self.assertEqual(230, len(slice_terms))
+        self.assertGreaterEqual(454, len(terms))
+        current = 23
+        for term in slice_terms:
+            self.assertEqual(term.accession, terms[current].accession)
+            current += 1
+        self.assertEqual(slice_terms[len(slice_terms) - 1], term_3)
+
+    def test_filters(self):
         """
         Test ontology terms api filtering options
         :return:
         """
         filters = {'short_form': 'EFO_0000405'}
         ontology = self.client.ontology('efo')
+        ontologies = self.client.ontologies(filters={'fake_filter': 'fake_value'})
+
         terms = ontology.terms(filters=filters)
         for term in terms:
             self.assertEqual(term.short_form, 'EFO_0000405')
@@ -165,7 +166,6 @@ class OntologyTestSuite(unittest.TestCase):
             self.assertEqual(term.obo_id, 'EFO:1000838')
             self.assertEqual(term.iri, 'http://www.ebi.ac.uk/efo/EFO_1000838')
 
-    @ignore_warnings
     def test_terms(self):
         """
         Test direct calls to terms entry point.
@@ -173,24 +173,11 @@ class OntologyTestSuite(unittest.TestCase):
         :return:
         """
         term_1 = helpers.Term(ontology_name='cco', iri='http://semanticscience.org/resource/SIO_010043')
-        ancestors = term_1.ancestors()
+        ancestors = term_1.load_relation('ancestors')
         for ancestor in ancestors:
             self._checkTerm(ancestor)
         self._checkTerm(term_1)
-        # actually load data from OLS
-        with warnings.catch_warnings(record=True) as w:
-            # Cause all warnings to always be triggered.
-            warnings.simplefilter("always")
-            # Trigger a warning.
-            term_2 = self.client.term(term_1.iri)
-            self.assertIsInstance(term_2, ListClientMixin)
-            self._checkTerms(term_2)
 
-            # Verify some things
-            self.assertEqual(len(w), 1)
-            self.assertTrue(issubclass(w[-1].category, UserWarning))
-
-    @ignore_warnings
     def test_dynamic_links(self):
         term = helpers.Term(ontology_name='so', iri='http://purl.obolibrary.org/obo/SO_0000104')
         for relation in term.relations_types:
@@ -201,7 +188,6 @@ class OntologyTestSuite(unittest.TestCase):
         term = helpers.Term(ontology_name='efo', iri='http://www.ebi.ac.uk/efo/EFO_0004799')
         self.assertIn('has_disease_location', term.relations_types)
 
-    @ignore_warnings
     def test_search(self):
         """
         Need more tests for filtering and selected fields in Results
@@ -238,7 +224,6 @@ class OntologyTestSuite(unittest.TestCase):
             clazz.append(mix.__class__.__name__) if mix.__class__.__name__ not in clazz else None
         self.assertGreater(len(clazz), 1)
 
-    @ignore_warnings
     def test_individuals(self):
         self.assertTrue(True)
         ontology = self.client.ontology('aero')
@@ -251,23 +236,52 @@ class OntologyTestSuite(unittest.TestCase):
             stop += 1
             # print(stop, indi)
 
+    def test_reverse_range(self):
+        ontologies = self.client.ontologies()
+        sliced = ontologies[158:110]
+        self.assertEqual(len(sliced), 48)
+        i = 158
+        for ontology in sliced:
+            self.assertEqual(ontology.ontology_id, ontologies[i].ontology_id)
+            i -= 1
 
-    @ignore_warnings
-    def testRangeTerms(self):
-        ontology = self.client.ontology('go')
-        terms = ontology.terms()
-        sliced = terms[10258:10630]
-        self.assertEqual(len(sliced), 372)
-        i = 10258
-        for term in sliced:
-            self.assertEqual(term.accession, terms[i].accession)
+        sliced = ontologies[185:210]
+        self.assertEqual(len(sliced), 25)
+        i = 185
+        for ontology in sliced:
+            self.assertEqual(ontology.ontology_id, ontologies[i].ontology_id)
             i += 1
 
-        # test now with page reload page inside
-        sliced = terms[8567:10630]
-        self.assertEqual(len(sliced), 2063)
-        i = 8567
-        for term in sliced:
-            self.assertEqual(term.accession, terms[i].accession)
-            i += 1
+    def test_error_search(self):
+        properties = self.client.search(query='goslim_yeast', filters={'ontology': 'go', 'type': 'property'})
+        for prop in properties:
+            details = self.client.detail(ontology_name='go', iri=prop.iri, type=helpers.Property)
+            print(details)
 
+    def test_accessions(self):
+        accessions = ['TopObjectProperty', 'SubsetProperty', 'ObsoleteClass']
+        for accession in accessions:
+            prop = helpers.Property(short_form=accession)
+            self.assertIsNone(prop.accession)
+        iris = ['http://purl.obolibrary.org/obo/go#gocheck_do_not_annotate']
+        for accession in iris:
+            prop = helpers.Property(iri=accession)
+            self.assertIsNotNone(prop.accession)
+            o_property = self.client.property(accession, unique=False)
+            self.assertIsNotNone(o_property)
+            self._checkProperties(o_property)
+
+    def test_exception_retry(self):
+        # unknown URI
+        with self.assertRaises(ebi.ols.api.exceptions.NotFoundException):
+            h_term = helpers.Term(ontology_name='so', iri='http://purl.obolibrary.org/obo/SO_99999999')
+            self.client.detail(h_term)
+        with self.assertRaises(ebi.ols.api.exceptions.NotFoundException):
+            self.client.ontology('unexisting_ontology')
+        with self.assertRaises(ebi.ols.api.exceptions.BadParameter):
+            filters = {'accession': 'EFO:0000405'}
+            self.client.terms(filters=filters)
+        with self.assertRaises(ebi.ols.api.exceptions.ObjectNotRetrievedError):
+            prop = helpers.Property(iri='http://purl.obolibrary.org/obo/uberon/insect-anatomy#efo_slim',
+                                    ontology_name='efo')
+            self.client.detail(prop)